@@ -9,7 +9,7 @@ import os
 import sys
 import logging
 import json
-from typing import Dict, Any, Optional, Union, BinaryIO, List
+from typing import Any, BinaryIO
 import openai
 from tenacity import retry, stop_after_attempt, wait_exponential, retry_if_exception_type
 
@@ -39,29 +39,48 @@ class AudioTranscriptionError(APIError):
     """Exception raised for audio transcription errors."""
     pass
 
+def _is_reasoning_model(model: str) -> bool:
+    """
+    Check if a model is a reasoning model (GPT-5, o1, o3 series).
+
+    Reasoning models use max_completion_tokens instead of max_tokens.
+
+    Args:
+        model: Model name to check
+
+    Returns:
+        True if model is a reasoning model
+    """
+    model_lower = model.lower()
+    return (
+        model_lower.startswith('gpt-5') or
+        model_lower.startswith('o1') or
+        model_lower.startswith('o3')
+    )
+
 class OpenAIClient:
     """
     Client for interacting with OpenAI's APIs with retry logic and error handling.
     """
-    
-    def __init__(self, api_key: Optional[str] = None, logger: Optional[logging.Logger] = None):
+
+    def __init__(self, api_key: str | None = None, logger: logging.Logger | None = None):
         """
         Initialize client with API key and logger.
-        
+
         Args:
             api_key: OpenAI API key (defaults to OPENAI_API_KEY environment variable)
             logger: Logger instance for output logging
-            
+
         Raises:
             ValueError: If no API key is provided and none is found in environment
         """
         self.api_key = api_key or os.environ.get('OPENAI_API_KEY')
         if not self.api_key:
             raise ValueError("OpenAI API key is required. Set OPENAI_API_KEY environment variable or pass as parameter.")
-            
+
         self.logger = logger or get_logger(__name__)
         self.client = openai.OpenAI(api_key=self.api_key)
-        
+
         # Configure OpenAI client's HTTP logger to use DEBUG level instead of INFO
         # Try multiple possible logger names used by different OpenAI library versions
         for logger_name in ["openai._client", "openai.http_client", "_client"]:
@@ -77,7 +96,7 @@ class OpenAIClient:
     )
     def transcribe_audio(
         self,
-        audio_file: Union[str, BinaryIO],
+        audio_file: str | BinaryIO,
         model: str = "whisper-1",
         prompt: str = "",
         language: str = 'en',
@@ -117,19 +136,19 @@ class OpenAIClient:
                 
         except openai.RateLimitError as e:
             self.logger.warning(f"Rate limit exceeded: {str(e)}")
-            raise APIRateLimitError(f"OpenAI API rate limit exceeded: {str(e)}")
-            
+            raise APIRateLimitError(f"OpenAI API rate limit exceeded: {str(e)}") from e
+
         except openai.AuthenticationError as e:
             self.logger.error(f"Authentication error: {str(e)}")
-            raise APIAuthenticationError(f"OpenAI API authentication error: {str(e)}")
-            
+            raise APIAuthenticationError(f"OpenAI API authentication error: {str(e)}") from e
+
         except openai.APIConnectionError as e:
             self.logger.warning(f"API connection error: {str(e)}")
-            raise APIConnectionError(f"OpenAI API connection error: {str(e)}")
-            
+            raise APIConnectionError(f"OpenAI API connection error: {str(e)}") from e
+
         except Exception as e:
             self.logger.error(f"Error transcribing audio: {str(e)}")
-            raise APIError(f"OpenAI API error: {str(e)}")
+            raise APIError(f"OpenAI API error: {str(e)}") from e
     
     def _transcribe_with_file(
         self,
@@ -182,68 +201,97 @@ class OpenAIClient:
         user_prompt: str,
         model: str = "gpt-4o",
         temperature: float = 0.0,
-        max_tokens: int = 1000
+        max_tokens: int = 1000,
+        reasoning_effort: str | None = None,
+        verbosity: str | None = None
     ) -> str:
         """
         Generate text using OpenAI's chat completion API.
-        
+
         Args:
             system_prompt: System message content
             user_prompt: User message content
-            model: Model to use (e.g., "gpt-4o", "gpt-4o-mini")
-            temperature: Temperature for generation
-            max_tokens: Maximum tokens to generate
-            
+            model: Model to use (e.g., "gpt-4o", "gpt-5", "gpt-5-mini")
+            temperature: Temperature for generation (ignored for GPT-5/reasoning models)
+            max_tokens: Maximum tokens to generate (max_completion_tokens for GPT-5)
+            reasoning_effort: Reasoning effort for GPT-5 models ("minimal", "low", "medium", "high")
+            verbosity: Verbosity for GPT-5 models ("low", "medium", "high")
+
         Returns:
             Generated text response
-            
+
         Raises:
             APIError: For API-related errors
+
+        Note:
+            GPT-5 and reasoning models (o1, o3) do not support custom temperature
+            (only default of 1) and use max_completion_tokens instead of max_tokens.
         """
         messages = [
             {"role": "system", "content": system_prompt},
             {"role": "user", "content": user_prompt}
         ]
-        
+
+        # Build API call parameters
+        api_params = {
+            "model": model,
+            "messages": messages,
+            "n": 1,
+            "stop": None
+        }
+
+        # GPT-5 and reasoning models use max_completion_tokens instead of max_tokens
+        # and do not support custom temperature (only default of 1)
+        if _is_reasoning_model(model):
+            api_params["max_completion_tokens"] = max_tokens
+            # Don't set temperature for reasoning models - they only support default (1)
+
+            # Add GPT-5 specific parameters if provided
+            if reasoning_effort:
+                api_params["reasoning_effort"] = reasoning_effort
+            if verbosity:
+                api_params["verbosity"] = verbosity
+        else:
+            api_params["max_tokens"] = max_tokens
+            api_params["temperature"] = temperature
+
         try:
-            response = self.client.chat.completions.create(
-                model=model,
-                messages=messages,
-                temperature=temperature,
-                max_tokens=max_tokens,
-                n=1,
-                stop=None
-            )
-            
+            response = self.client.chat.completions.create(**api_params)
+
             self.logger.debug(f"HTTP Request: POST https://api.openai.com/v1/chat/completions \"HTTP/1.1 200 OK\"")
-            
+
             if not response.choices or not response.choices[0].message.content.strip():
-                self.logger.warning(f"Empty response from API: user_prompt='{user_prompt[:128]}...'")
+                # Empty responses can happen with reasoning models (GPT-5, o1, o3)
+                # Use debug level for reasoning models since they have temperature=1 (less deterministic)
+                if _is_reasoning_model(model):
+                    self.logger.debug(f"Empty response from reasoning model: user_prompt='{user_prompt[:128]}...'")
+                else:
+                    self.logger.warning(f"Empty response from API: user_prompt='{user_prompt[:128]}...'")
                 return ""
-                
+
             return response.choices[0].message.content.strip()
             
         except openai.RateLimitError as e:
             self.logger.warning(f"Rate limit exceeded: {str(e)}")
-            raise APIRateLimitError(f"OpenAI API rate limit exceeded: {str(e)}")
-            
+            raise APIRateLimitError(f"OpenAI API rate limit exceeded: {str(e)}") from e
+
         except openai.AuthenticationError as e:
             self.logger.error(f"Authentication error: {str(e)}")
-            raise APIAuthenticationError(f"OpenAI API authentication error: {str(e)}")
-            
+            raise APIAuthenticationError(f"OpenAI API authentication error: {str(e)}") from e
+
         except openai.APIConnectionError as e:
             self.logger.warning(f"API connection error: {str(e)}")
-            raise APIConnectionError(f"OpenAI API connection error: {str(e)}")
-            
+            raise APIConnectionError(f"OpenAI API connection error: {str(e)}") from e
+
         except Exception as e:
             self.logger.error(f"Error in chat completion: {str(e)}")
-            raise APIError(f"OpenAI API error: {str(e)}")
+            raise APIError(f"OpenAI API error: {str(e)}") from e
 
-# Global client instance for wrapper functions  
-_global_client: Optional[OpenAIClient] = None
-openai_client: Optional[OpenAIClient] = None  # For test compatibility
+# Global client instance for wrapper functions
+_global_client: OpenAIClient | None = None
+openai_client: OpenAIClient | None = None  # For test compatibility
 
-def get_openai_client(api_key: Optional[str] = None) -> OpenAIClient:
+def get_openai_client(api_key: str | None = None) -> OpenAIClient:
     """
     Get OpenAI client instance (creates global instance if needed).
     
@@ -259,38 +307,46 @@ def get_openai_client(api_key: Optional[str] = None) -> OpenAIClient:
         openai_client = _global_client  # For test compatibility
     return _global_client
 
-def call_llm(user_prompt: str, system_prompt: Optional[str] = None, model: str = "gpt-4o", 
-             temperature: float = 0.1, max_tokens: Optional[int] = None) -> str:
+def call_llm(user_prompt: str, system_prompt: str | None = None, model: str = "gpt-4o",
+             temperature: float = 0.1, max_tokens: int | None = None,
+             reasoning_effort: str | None = None, verbosity: str | None = None) -> str:
     """
     Call LLM with user and optional system prompts - the authoritative LLM interface.
-    
+
     This is the single point of entry for all LLM calls throughout the codebase.
     It provides consistent error handling, retry logic, and parameter standardization.
-    
+
     Args:
         user_prompt: User input text to process
         system_prompt: Optional system prompt to set context (defaults to None)
         model: Model to use for completion (defaults to "gpt-4o")
-        temperature: Temperature for generation (defaults to 0.1)
+        temperature: Temperature for generation (defaults to 0.1, ignored for GPT-5/reasoning models)
         max_tokens: Maximum tokens to generate (defaults to None for model default)
-        
+        reasoning_effort: Reasoning effort for GPT-5 models ("minimal", "low", "medium", "high")
+        verbosity: Verbosity for GPT-5 models ("low", "medium", "high")
+
     Returns:
         Generated response text
-        
+
     Raises:
         EmptyResponseError: If response is empty
         APIError: For other API errors
-        
+
     Note:
-        This function supports both new parameter order (user_prompt first) and 
+        This function supports both new parameter order (user_prompt first) and
         legacy parameter order (system_prompt first) for backward compatibility.
+
+        GPT-5 and reasoning models (o1, o3 series):
+        - Use max_completion_tokens instead of max_tokens
+        - Do not support custom temperature (only default of 1)
+        - Support reasoning_effort and verbosity parameters
     """
     # Backward compatibility: detect if old parameter order is being used
     # Legacy calls might pass system_prompt as first parameter
     if isinstance(user_prompt, str) and system_prompt is None and "system" in user_prompt.lower()[:100]:
         # Likely legacy call pattern - but this is fragile, so we'll trust the new order
         pass
-    
+
     # Use global openai_client if available (for testing), otherwise get a new one
     global openai_client
     if openai_client is not None and hasattr(openai_client, 'chat'):
@@ -299,29 +355,45 @@ def call_llm(user_prompt: str, system_prompt: Optional[str] = None, model: str =
         if system_prompt:
             messages.append({"role": "system", "content": system_prompt})
         messages.append({"role": "user", "content": user_prompt})
-        
+
+        # Build API parameters
+        api_params = {
+            "model": model,
+            "messages": messages
+        }
+
+        # Handle GPT-5 models - use max_completion_tokens instead of max_tokens
+        # and do not support custom temperature (only default of 1)
+        if _is_reasoning_model(model):
+            # Don't set temperature for reasoning models - they only support default (1)
+            if max_tokens is not None:
+                api_params["max_completion_tokens"] = max_tokens
+            if reasoning_effort:
+                api_params["reasoning_effort"] = reasoning_effort
+            if verbosity:
+                api_params["verbosity"] = verbosity
+        else:
+            api_params["temperature"] = temperature
+            if max_tokens is not None:
+                api_params["max_tokens"] = max_tokens
+
         try:
-            response_obj = openai_client.chat.completions.create(
-                model=model,
-                messages=messages,
-                temperature=temperature,
-                max_tokens=max_tokens
-            )
-            
+            response_obj = openai_client.chat.completions.create(**api_params)
+
             if not response_obj.choices:
                 raise EmptyResponseError("No choices in response")
-            
+
             content = response_obj.choices[0].message.content
             if not content:
                 response = ""
             else:
                 response = content.strip()
-                
+
         except Exception as e:
             if "choices" in str(e).lower() or "empty" in str(e).lower():
-                raise EmptyResponseError(f"Empty response from API: {str(e)}")
+                raise EmptyResponseError(f"Empty response from API: {str(e)}") from e
             else:
-                raise APIError(f"API error: {str(e)}")
+                raise APIError(f"API error: {str(e)}") from e
     else:
         # For production - use our OpenAIClient wrapper
         client = get_openai_client()
@@ -330,12 +402,14 @@ def call_llm(user_prompt: str, system_prompt: Optional[str] = None, model: str =
             system_prompt=system_prompt,
             model=model,
             temperature=temperature,
-            max_tokens=max_tokens
+            max_tokens=max_tokens,
+            reasoning_effort=reasoning_effort,
+            verbosity=verbosity
         )
-    
+
     return response
 
-def transcribe_audio(audio_file: Union[str, BinaryIO], model: str = "whisper-1",
+def transcribe_audio(audio_file: str | BinaryIO, model: str = "whisper-1",
                     prompt: str = "", language: str = 'en',
                     response_format: str = "text", temperature: float = 0.05) -> Any:
     """
@@ -386,11 +460,11 @@ def transcribe_audio(audio_file: Union[str, BinaryIO], model: str = "whisper-1",
                 temperature=temperature
             )
     except APIError as e:
-        raise AudioTranscriptionError(f"Audio transcription failed: {str(e)}")
+        raise AudioTranscriptionError(f"Audio transcription failed: {str(e)}") from e
 
 # Internal implementation functions for testing
-def _call_llm_impl(system_prompt: str, user_prompt: str, model: str = "gpt-4o", 
-                   temperature: float = 0.1, max_tokens: Optional[int] = None) -> str:
+def _call_llm_impl(system_prompt: str, user_prompt: str, model: str = "gpt-4o",
+                   temperature: float = 0.1, max_tokens: int | None = None) -> str:
     """
     Internal implementation function for LLM calls (used by tests).
     
@@ -434,9 +508,9 @@ def _call_llm_impl(system_prompt: str, user_prompt: str, model: str = "gpt-4o",
         return content
         
     except openai.APIError as e:
-        raise APIError(f"OpenAI API error: {str(e)}")
+        raise APIError(f"OpenAI API error: {str(e)}") from e
 
-def _transcribe_audio_impl(audio_file: Union[str, BinaryIO], model: str = "whisper-1",
+def _transcribe_audio_impl(audio_file: str | BinaryIO, model: str = "whisper-1",
                           prompt: str = "", language: str = 'en',
                           response_format: str = "text", temperature: float = 0.05) -> Any:
     """
@@ -488,6 +562,6 @@ def _transcribe_audio_impl(audio_file: Union[str, BinaryIO], model: str = "whisp
         return transcription
         
     except openai.APIError as e:
-        raise APIError(f"OpenAI API error: {str(e)}")
+        raise APIError(f"OpenAI API error: {str(e)}") from e
 
 #fin
